Fix build errors batch 8: PanelWireReconcile, WireAnnotation, ViewStylePackApplier
- PanelWireReconcileCommand: fix long-to-int implicit conversion with explicit cast
- WireAnnotationCommands: fix fill variable use-before-declaration and TickStyle reference
- ViewStylePackApplier: add ApplyWorksetVisibility stub method

+++ StingTools/Commands/Electrical/PanelWireReconcileCommand.cs
@@ -327,7 +327,7 @@ namespace StingTools.Commands.Electrical
                 (int)BuiltInCategory.OST_LightingFixtures,
             };
 
-            int id = catId.Value;
+            int id = (int)catId.Value;
             foreach (int c in electricalCategories)
                 if (c == id) return true;
             return false;
